routes get, post, put, delete from messages

+++ controllers/messages.js
@@ -1,4 +1,5 @@
 const { getConnection } = require('../db');
+const { processAndSaveImage, deleteImage } = require('../helpers');
 
 // Rutas para anonimous user
 
@@ -66,20 +67,40 @@ async function getMessage(req, res, next) {
 async function newMessage(req, res, next) {
   try {
     const connection = await getConnection();
-    const { from_users_id, to_users_id, title, text, type, category } = req.body;
+    const { from_users_id, to_users_id, title, text, type, category, image } = req.body;
+
+    console.log(req.files.image);
+
+    if (!from_users_id || !to_users_id || !title || !text || !type || !category) {
+      const error = new Error('Los campos de, para, titulo, texto, tipo y categoria son obligatorios');
+      error.httpCode = 400;
+      throw error;
+    }
+
+    let savedFileName;
+
+    if (req.files && req.files.image) {
+      try {
+        savedFileName = await processAndSaveImage(req.files.image);
+      } catch (error) {
+        const imageError = new Error('No se ha podido procesar la imagen');
+        imageError.httpCode = 400;
+        throw imageError;
+      }
+    }
 
     await connection.query(
       `
-      insert into messages (from_users_id, to_users_id, title, text, type, category)
-      values (?, ?, ?, ?, ?, ?)`,
-      [from_users_id, to_users_id, title, text, type, category]
+      insert into messages (from_users_id, to_users_id, title, text, type, category, image)
+      values (?, ?, ?, ?, ?, ?, ?)`,
+      [from_users_id, to_users_id, title, text, type, category, savedFileName]
     );
 
     connection.release();
 
     res.send({
       status: 'ok',
-      data: { from_users_id, to_users_id, title, text, type, category },
+      data: { from_users_id, to_users_id, title, text, type, category, savedFileName },
     });
   } catch (error) {
     next(error);
@@ -88,6 +109,73 @@ async function newMessage(req, res, next) {
 
 // PUT - /messages/:id
 
+async function editMessage(req, res, next) {
+  try {
+    const connection = await getConnection();
+    const { id } = req.params;
+    const { title, text, type, category } = req.body;
+
+    if (!title || !text || !type || !category) {
+      const error = new Error('Los campos titulo, texto, type, category son obligatorios');
+      error.httpCode = 400;
+      throw error;
+    }
+
+    const [current] = await connection.query(`select image from messages where id=?`, [id]);
+    console.log(current);
+
+    if (!current.length) {
+      const error = new Error(`No existe el mensaje con id ${id}`);
+      error.httpCode = 400;
+      throw error;
+    }
+
+    let savedFileName;
+
+    console.log(req.files.image);
+
+    if (req.files && req.files.image) {
+      try {
+        savedFileName = await processAndSaveImage(req.files.image);
+
+        if (current[0] && current[0].image) {
+          await deleteImage(current[0].image);
+        }
+      } catch (error) {
+        const imageError = new Error('No se ha podido procesar la imagen');
+        imageError.httpCode = 400;
+        throw imageError;
+      }
+    } else {
+      savedFileName = current[0].image;
+    }
+
+    await connection.query(`update messages set title=?, text=?, type=?, category=?, image=? where id=?`, [
+      title,
+      text,
+      type,
+      category,
+      savedFileName,
+      id,
+    ]);
+
+    connection.release();
+
+    res.send({
+      status: 'ok',
+      data: {
+        title: title,
+        text: text,
+        type: type,
+        category: category,
+        image: savedFileName,
+      },
+    });
+  } catch (error) {
+    next(error);
+  }
+}
+
 // DELETE - /messages/:id
 async function deleteMessage(req, res, next) {
   try {
@@ -95,7 +183,25 @@ async function deleteMessage(req, res, next) {
 
     const connection = await getConnection();
 
+    const [current] = await connection.query(`select image from messages where id=?`, [id]);
+
+    if (!current.length) {
+      const error = new Error(`No existe el mensaje con id ${id}`);
+      error.httpCode = 400;
+      throw error;
+    }
+
+    if (current[0].image) {
+      await deleteImage(current[0].image);
+    }
+
     await connection.query(`delete from messages where id=?`, [id]);
+
+    // if (result.affectedRows === 0) {
+    //   const error = new Error(`El mensaje con la id: ${id} no existe`);
+    //   error.httpCode = 400;
+    //   throw error;
+    // }
 
     connection.release();
 
@@ -112,5 +218,6 @@ module.exports = {
   listMessages,
   getMessage,
   newMessage,
+  editMessage,
   deleteMessage,
 };
